+++ blocks/FileItem/FileItem.js
@@ -214,10 +214,13 @@ export class FileItem extends UploaderBlock {
     const mimeType = entry.getValue('mimeType');
     const fileName = entry.getValue('fileName');
 
-    const needMimeCheck = !!mimeType;
-    const needExtCheck = !!fileName;
-    const mimeOk = needMimeCheck ? matchMimeType(mimeType, allowedFileTypes) : true;
-    const extOk = needExtCheck ? matchExtension(fileName, allowedFileTypes) : true;
+    if (!mimeType || !fileName) {
+      // Skip client validation if mime type or file name are not available for some reasons
+      return;
+    }
+
+    const mimeOk = matchMimeType(mimeType, allowedFileTypes);
+    const extOk = matchExtension(fileName, allowedFileTypes);
 
     if (!mimeOk && !extOk) {
       // Assume file type is not allowed if both mime and ext checks fail
@@ -462,7 +465,7 @@ export class FileItem extends UploaderBlock {
         fileName: fileInfo.originalFilename,
         fileSize: fileInfo.size,
         isImage: fileInfo.isImage,
-        mimeType: fileInfo.mimeType,
+        mimeType: fileInfo.contentInfo?.mime?.mime ?? fileInfo.mimeType,
         uuid: fileInfo.uuid,
         cdnUrl: fileInfo.cdnUrl,
       });